Move SubscriptionDetails to its own class
