+++ app/services/bulk_unsubscribe_service.rb
@@ -117,35 +117,7 @@ module BulkUnsubscribeService
       - [<%= details.replacement_title %>](<%= add_utm(details.replacement_url, utm_parameters) %>)
     <% end %>
   BODY
-
-  class SubscriptionDetails
-
-    def initialize(subscription, replacement)
-      @subscription = subscription
-      @replacement = replacement
-    end
-
-    def subscriber_list
-      @_subscriber_list = @subscription.subscriber_list
-    end
-
-    def title
-      subscriber_list.title
-    end
-
-    def links
-      subscriber_list.links
-    end
-
-    def replacement_title
-      @replacement.title
-    end
-
-    def replacement_url
-      @replacement.url
-    end
-
-  end
+  
 end
 
 # rubocop:enable Metrics/BlockLength
